Handle response objects
When the script returns a table/struct it will be handled on the Go side
with the relevant headers and status code being set. Like the request
object this follows the Circlet style.

Also improved the module loader predicate to actually validate that the
module being loaded is one provided by Spinnerette.