Reorganize and Cleanup project
Moved things around so that the Janet bindings live in their own folder.
The idea is that everything C related lives in bindings and everything
Janet related is in libs

Also pulled in the C libraries for Janet JSON and Sqlite3. They aren't
hooked up yet and there are some issues that I need to solve involving
that first.