Wrap and bind native libraries
Added bindings to the two native libraries that I pulled in earlier. So
now Spinnerette has JSON and SQLite3 support! Both use the official
Janet libraries without modifications.

This involved some odd abuse of the C preprocessor and a bit of
fiddling, but it all seems to be working nicely.

This addition also improved the module loader somewhat, which is now a
Go function instead of a generated Janet function. More to come on this
as the loader improves.