Fixup the Module loader with startup.janet
Introduced a new fine, startup.janet, which is evaluated when the VM is
first started and can be used to set things up from there.
In that I added some new machinery for a functional module loader.
Took quite a bit of fiddling, but it seems to be working nicely.